+++ client/rtype/rtype/AProjectile.hpp
@@ -15,14 +15,14 @@ public:
 	virtual void collision(IClient *client, AEntity *other) = 0;
 	virtual void applyCollision(CollisionType type) = 0;
 
-	void setOwnerID(uint8_t ownerID);
-	uint8_t getOwnerID(void) const;
+	void setOwnerID(uint16_t ownerID);
+	uint16_t getOwnerID(void) const;
 
 	void setLevel(uint8_t level);
 	uint8_t getLevel(void) const;
 
 private:
-	uint8_t _ownerID;
+	uint16_t _ownerID;
 	uint8_t _currentLevel;
 };
